Moved trivial blob backend impl to private header
