+++ include/soci/sqlite3/soci-sqlite3.h
@@ -20,6 +20,7 @@
 #include <cstdint>
 #include <vector>
 #include <soci/soci-backend.h>
+#include <private/soci-trivial-blob-backend.h>
 
 // Disable flood of nonsense warnings generated for SQLite
 #ifdef _MSC_VER
